refactor: update accessibility settings and return value
